+++ src/ApolloClient.ts
@@ -16,6 +16,11 @@ import {
 } from 'graphql';
 
 import {
+  HeuristicFragmentMatcher,
+  FragmentMatcherInterface,
+} from './data/fragmentMatcher';
+
+import {
   createApolloStore,
   ApolloStore,
   createApolloReducer,
@@ -79,6 +84,7 @@ import {
   version,
 } from './version';
 
+
 /**
  * This type defines a "selector" function that receives state from the Redux store
  * and returns the part of it that is managed by ApolloClient
@@ -92,6 +98,20 @@ const DEFAULT_REDUX_ROOT_KEY = 'apollo';
 function defaultReduxRootSelector(state: any) {
   return state[DEFAULT_REDUX_ROOT_KEY];
 }
+
+function defaultDataIdFromObject (result: any): string | null {
+  if (result.__typename) {
+    if (result.id !== undefined) {
+      return `${result.__typename}:${result.id}`;
+    }
+    if (result._id !== undefined) {
+      return `${result.__typename}:${result._id}`;
+    }
+  }
+  return null;
+}
+
+let hasSuggestedDevtools = false;
 
 /**
  * This is the primary Apollo Client class. It is used to send GraphQL documents (i.e. queries
@@ -115,6 +135,7 @@ export default class ApolloClient implements DataProxy {
 
   private devToolsHookCb: Function;
   private proxy: DataProxy | undefined;
+  private fragmentMatcher: FragmentMatcherInterface;
 
   /**
    * Constructs an instance of {@link ApolloClient}.
@@ -142,6 +163,7 @@ export default class ApolloClient implements DataProxy {
    * @param queryDeduplication If set to false, a query will still be sent to the server even if a query
    * with identical parameters (query, variables, operationName) is already in flight.
    *
+   * @param fragmentMatcher A function to use for matching fragment conditions in GraphQL documents
    */
 
   constructor(options: {
@@ -155,17 +177,21 @@ export default class ApolloClient implements DataProxy {
     customResolvers?: CustomResolverMap,
     connectToDevTools?: boolean,
     queryDeduplication?: boolean,
+    fragmentMatcher?: FragmentMatcherInterface,
   } = {}) {
+    let {
+      dataIdFromObject,
+    } = options;
     const {
       networkInterface,
       reduxRootSelector,
       initialState,
-      dataIdFromObject,
       ssrMode = false,
       ssrForceFetchDelay = 0,
       addTypename = true,
       customResolvers,
       connectToDevTools,
+      fragmentMatcher,
       queryDeduplication = true,
     } = options;
 
@@ -175,12 +201,18 @@ export default class ApolloClient implements DataProxy {
       throw new Error('"reduxRootSelector" must be a function.');
     }
 
+    if (typeof fragmentMatcher === 'undefined') {
+      this.fragmentMatcher = new HeuristicFragmentMatcher();
+    } else {
+      this.fragmentMatcher = fragmentMatcher;
+    }
+
     this.initialState = initialState ? initialState : {};
     this.networkInterface = networkInterface ? networkInterface :
       createNetworkInterface({ uri: '/graphql' });
     this.addTypename = addTypename;
     this.disableNetworkFetches = ssrMode || ssrForceFetchDelay > 0;
-    this.dataId = dataIdFromObject;
+    this.dataId = dataIdFromObject = dataIdFromObject || defaultDataIdFromObject;
     this.fieldWithArgs = storeKeyNameFromFieldNameAndArgs;
     this.queryDeduplication = queryDeduplication;
 
@@ -207,6 +239,26 @@ export default class ApolloClient implements DataProxy {
 
     if (typeof connectToDevTools === 'undefined' ? defaultConnectToDevTools : connectToDevTools) {
       (window as any).__APOLLO_CLIENT__ = this;
+    }
+
+    /**
+     * Suggest installing the devtools for developers who don't have them
+     */
+    if (!hasSuggestedDevtools && !isProduction()) {
+      hasSuggestedDevtools = true;
+      if ( typeof window !== 'undefined' && window.document && window.top === window.self) {
+
+        // First check if devtools is not installed
+        if (typeof (window as any).__APOLLO_DEVTOOLS_GLOBAL_HOOK__ === 'undefined') {
+          // Only for Chrome
+          if (navigator.userAgent.indexOf('Chrome') > -1) {
+            // tslint:disable-next-line
+            console.debug('Download the Apollo DevTools ' +
+            'for a better development experience: ' +
+            'https://chrome.google.com/webstore/detail/apollo-client-developer-t/jdkknkkbebbapilgoeccciglkfbmbnfm');
+          }
+        }
+      }
     }
 
     this.version = version;
@@ -467,6 +519,7 @@ export default class ApolloClient implements DataProxy {
       addTypename: this.addTypename,
       reducerConfig: this.reducerConfig,
       queryDeduplication: this.queryDeduplication,
+      fragmentMatcher: this.fragmentMatcher,
     });
   };
 
@@ -481,6 +534,8 @@ export default class ApolloClient implements DataProxy {
       this.proxy = new ReduxDataProxy(
         this.store,
         this.reduxRootSelector || defaultReduxRootSelector,
+        this.fragmentMatcher,
+        this.reducerConfig,
       );
     }
     return this.proxy;