+++ test/tests.ts
@@ -18,13 +18,14 @@ require('source-map-support').install();
 console.warn = console.error = (...messages: string[]) => {
   console.log(`==> Error in test: Tried to log warning or error with message:
 `, ...messages);
-  if (!process.env.CI) {
+  if ( (!process.env.CI) && (!process.env.COV) ) {
     process.exit(1);
   }
 };
 
 process.on('unhandledRejection', () => {});
 
+import './fragmentMatcher';
 import './writeToStore';
 import './readFromStore';
 import './roundtrip';