+++ typings.d.ts
@@ -3,7 +3,7 @@
   GRAPHQL
 
 */
-declare module 'graphql-tag/parser' {
+declare module 'graphql/language/parser' {
   import { Source, ParseOptions, DocumentNode } from 'graphql';
   // XXX figure out how to directly export this method
   function parse(
@@ -11,3 +11,8 @@ declare module 'graphql-tag/parser' {
       options?: ParseOptions
   ): DocumentNode;
 }
+
+declare module 'graphql/language/printer' {
+  function print(ast: any): string;
+}
+